requirements.txt
flask
ultralytics
torch        # CPU or torch-cuXX if you have a GPU locally
opencv-python
pytesseract  # or easyocr if you prefer
pyserial     # for Arduino comms